refactor(plugins): Move plugin module loading into case-entities
Extracts the name validation, working-directory resolution and dynamic
import (with its fallbacks) from BoundaryPluginLoader into a reusable case-entities-internal
loadPluginModuleContents helper, so that the DSL generator can load
plugin modules the same way the connector does. No behaviour change.
