+++ packages/case-core/src/connectors/loadPlugins/index.ts
@@ -1,2 +1,1 @@
 export * from './PluginLoader';
-export * from './resolve';
