+++ packages/case-entities/package.json
@@ -31,6 +31,7 @@
   },
   "scripts": {
     "build": "rimraf dist && tsc --project tsconfig.build.json",
+    "test": "jest",
     "lint": "eslint",
     "lint:fix": "eslint --fix",
     "format:base": "prettier --parser typescript \"{src,bin}/*.ts\"",
@@ -68,6 +69,7 @@
   "dependencies": {
     "@contract-case/case-core-plugin-http-dsl": "0.30.1",
     "@contract-case/case-core-plugin-function-dsl": "0.30.1",
+    "@contract-case/case-plugin-base": "0.30.1",
     "@contract-case/case-plugin-dsl-types": "0.30.1"
   }
 }
