+++ packages/case-entities/src/index.ts
@@ -7,3 +7,4 @@
 export * from './matchers';
 export * from './mocks';
 export * from './states';
+export * from './plugins';
