+++ packages/case-entities/src/plugins/index.ts
@@ -1,0 +1,2 @@
+export * from './loadModule';
+export * from './resolve';
