+++ packages/case-entities/src/plugins/resolve.spec.ts
