+++ packages/case-entities/src/plugins/resolve.ts
